% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_decomp.R
\name{.shape_term_names}
\alias{.shape_term_names}
\title{为 shape 系数生成命名（内部）}
\usage{
.shape_term_names(degree)
}
\arguments{
\item{degree}{阶数}
}
\value{
字符向量
}
\description{
为 shape 系数生成命名（内部）
}
\keyword{internal}
